[REFACTOR] post 테스트 분리
