+++ src/tests/units/components/feature/TimeSetting.test.tsx
@@ -2,97 +2,10 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import Postpage from "@/app/post/create/[...type]/page";
-import Provider from "@/components/provider/Provider";
-import { renderLoggedInLayout } from "@/utils/util-test";
-import TimeSetting from "@/components/post/discussionTimeSetting/TimeSetting";
 import CustomNumberInput from "@/components/post/discussionTimeSetting/CustomNumberInput";
 import useInput from "@/hooks/useInput";
-import Calendar from "@/components/post/discussionTimeSetting/calendar/Calendar";
-import { CalendarManager } from "@/utils/util-date";
-
-const today = new Date();
-const calendarManager = new CalendarManager(
-  today.getFullYear(),
-  today.getMonth(),
-);
-
-test("post페이지 렌더링 테스트", async () => {
-  const { unmount } = render(<Postpage params={{ type: ["tech"] }} />, {
-    wrapper: Provider,
-  });
-
-  expect(screen.queryByText("토의 시간 설정")).not.toBeInTheDocument();
-
-  unmount();
-  render(<Postpage params={{ type: ["discussion"] }} />, { wrapper: Provider });
-
-  expect(await screen.findByText("토의 시간 설정")).toBeInTheDocument();
-
-  expect(
-    await screen.findByPlaceholderText("내용을 입력해주세요."),
-  ).toBeInTheDocument();
-});
-
-test("토의 시간 설정 모달 렌더링 테스트", async () => {
-  render(<TimeSetting setTimeSetting={jest.fn} />, { wrapper: Provider });
-  await userEvent.click(await screen.findByText("토의 시간 설정"));
-  expect(await screen.findByText("시작 시간")).toBeInTheDocument();
-  expect(await screen.findByText("종료 시간")).toBeInTheDocument();
-  expect(await screen.findByText("기간")).toBeInTheDocument();
-  expect(await screen.findAllByDisplayValue("오전")).toHaveLength(2);
-  expect(await screen.findAllByDisplayValue("12시")).toHaveLength(2);
-  expect(await screen.findAllByDisplayValue("00분")).toHaveLength(2);
-  expect(await screen.findByDisplayValue(/월/)).toBeInTheDocument();
-  expect(await screen.findByDisplayValue(/일/)).toBeInTheDocument();
-});
-
-describe("태그 인풋 테스트", () => {
-  beforeEach(() => {
-    renderLoggedInLayout(<Postpage params={{ type: ["tech"] }} />, {
-      withHeader: false,
-    });
-  });
-  const modalText = "해쉬태그 또는 엔터를 입력하여 태그를 등록할 수 있습니다.";
-  const tagInputPlaceholder = "태그는 최대 3개까지 입력할 수 있습니다.";
-
-  test("태그 인풋 클릭 시 안내 문구가 뜨고 다른 곳 클릭하면 사라져야 함", async () => {
-    expect(screen.queryByText(modalText)).not.toBeInTheDocument();
-    await userEvent.click(
-      await screen.findByPlaceholderText(tagInputPlaceholder),
-    );
-    expect(await screen.findByText(modalText)).toBeInTheDocument();
-    await userEvent.click(
-      await screen.findByPlaceholderText("제목을 입력해주세요."),
-    );
-    expect(screen.queryByText(modalText)).not.toBeInTheDocument();
-  });
-
-  test("태그 3개까지만 추가되는지 테스트", async () => {
-    const input = await screen.findByPlaceholderText(tagInputPlaceholder);
-    await userEvent.type(input, "리액트");
-    await userEvent.keyboard("{enter}");
-    expect(await screen.findByText("#리액트")).toBeInTheDocument();
-    await userEvent.type(input, "넥스트");
-    await userEvent.keyboard("{enter}");
-    expect(await screen.findByText("#넥스트")).toBeInTheDocument();
-    await userEvent.type(input, "타입스크립트");
-    await userEvent.keyboard("{enter}");
-    expect(await screen.findByText("#타입스크립트")).toBeInTheDocument();
-    await userEvent.type(input, "자바스크립트");
-    await userEvent.keyboard("{enter}");
-    expect(screen.queryByText("#자바스크립트")).not.toBeInTheDocument();
-  });
-
-  test("태그 잘 삭제되는지 테스트", async () => {
-    const input = await screen.findByPlaceholderText(tagInputPlaceholder);
-    await userEvent.type(input, "리액트");
-    await userEvent.keyboard("{enter}");
-    expect(await screen.findByText("#리액트")).toBeInTheDocument();
-    await userEvent.click(await screen.findByAltText("default_close"));
-    expect(screen.queryByText("#리액트")).not.toBeInTheDocument();
-  });
-});
+import Provider from "@/components/provider/Provider";
+import TimeSetting from "@/components/post/discussionTimeSetting/TimeSetting";
 
 describe("토의 시간 설정 테스트", () => {
   const TestComponent: React.FC<{
@@ -298,91 +211,5 @@ describe("토의 시간 설정 테스트", () => {
       await userEvent.click(await screen.findByText("기간"));
       expect(await screen.findByDisplayValue("28일")).toBeInTheDocument();
     });
-  });
-});
-
-describe("달력 테스트", () => {
-  beforeEach(() => {
-    render(<CalendarTestComponent />, { wrapper: Provider });
-  });
-
-  const CalendarTestComponent = () => {
-    const year = useInput(today.getFullYear());
-    const month = useInput(today.getMonth());
-    const date = useInput(today.getDate());
-    return (
-      <Calendar
-        yearValue={+year.value}
-        monthValue={+month.value}
-        dateValue={+date.value}
-        setYearValue={year.setValue}
-        setMonthValue={month.setValue}
-        setDateValue={date.setValue}
-      />
-    );
-  };
-
-  test("이전 달 일을 누르면 이전 달로 이동해야 함 / 과거날짜 누르면 알럿이 떠야 함", async () => {
-    const prevMonthLastDate = calendarManager.getPrevMonthLastDate();
-    const prevMonthLastDay = await screen.findByTestId(
-      `prevMonthDay_${prevMonthLastDate}`,
-    );
-    await userEvent.click(prevMonthLastDay);
-    expect(
-      await screen.findByText(`. ${today.getMonth() - 1}`, { exact: false }),
-    ).toBeInTheDocument();
-
-    await userEvent.click(await screen.findByText("15"));
-    expect(window.alert).toHaveBeenCalledWith("이미 지난 날짜입니다.");
-  });
-
-  test("다음 달 일을 누르면 다음 달로 이동해야 함", async () => {
-    await userEvent.click(await screen.findByTestId(`nextMonthDay_1`));
-    expect(
-      await screen.findByText(`. ${today.getMonth() + 1}`, { exact: false }),
-    ).toBeInTheDocument();
-  });
-
-  test("달 이동 테스트", async () => {
-    const nextButton = await screen.findByTestId("next_month");
-    const prevButton = await screen.findByTestId("prev_month");
-    await userEvent.click(nextButton);
-    expect(
-      await screen.findByText(`. ${today.getMonth() + 1}`, { exact: false }),
-    ).toBeInTheDocument();
-    await userEvent.click(prevButton);
-    expect(
-      await screen.findByText(`. ${today.getMonth()}`, { exact: false }),
-    ).toBeInTheDocument();
-  });
-});
-
-describe("달력 날짜 선택 테스트", () => {
-  beforeEach(async () => {
-    render(<TimeSetting setTimeSetting={jest.fn} />, { wrapper: Provider });
-    await userEvent.click(await screen.findByText("토의 시간 설정"));
-    await userEvent.click(await screen.findByTestId("calendar"));
-  });
-
-  test("달력을 열면 오늘 날짜가 선택되어 있어야 함", async () => {
-    expect(
-      await screen.findByTestId(`currentMonthDay_${today.getDate()}`),
-    ).toHaveClass("bg-[#967AC3]");
-  });
-
-  test("달력에서 날짜를 선택하면 달력이 사라지고 인풋에 반영되어야 함", async () => {
-    await userEvent.click(await screen.findByTestId(`nextMonthDay_1`));
-    await userEvent.click(await screen.findByTestId(`currentMonthDay_1`));
-    expect(screen.queryByTestId(`currentMonthDay_1`)).not.toBeInTheDocument();
-    expect(await screen.findByDisplayValue("01일")).toBeInTheDocument();
-  });
-
-  test("달력의 선택된 날짜는 색이 바뀌어야 함", async () => {
-    await userEvent.click(await screen.findByTestId(`nextMonthDay_1`));
-    await userEvent.click(await screen.findByTestId(`currentMonthDay_1`));
-    await userEvent.click(await screen.findByTestId("calendar"));
-    expect(await screen.findByTestId(`currentMonthDay_1`)).toHaveClass(
-      "bg-[#967AC3]",
-    );
   });
 });
